課題グループをエンティティに変更
