+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/domain/exercise/entity/exercise-group.ts
@@ -1,0 +1,28 @@
+import { Entity } from "domain/shared/entity";
+import { Identifier } from "domain/shared/identifier";
+
+export interface IExerciseGroup {
+  name: string;
+}
+
+export class ExerciseGroup extends Entity<IExerciseGroup> {
+  public get name(): string {
+    return this.props.name;
+  }
+
+  public static create(props: IExerciseGroup): ExerciseGroup {
+    if (props.name === "") {
+      throw new Error("Illegal name value.");
+    }
+
+    return new ExerciseGroup(props);
+  }
+
+  public static rebuild(id: Identifier, props: IExerciseGroup): ExerciseGroup {
+    if (props.name === "") {
+      throw new Error("Illegal name value.");
+    }
+
+    return new ExerciseGroup(props, id);
+  }
+}
