+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/domain/exercise/entity/exercise.ts
@@ -1,4 +1,4 @@
-import { ExerciseGroup } from "domain/exercise/value-object/exercise-group";
+import { ExerciseGroup } from "domain/exercise/entity/exercise-group";
 import { AggregateRoot } from "domain/shared/aggregate-root";
 import { Identifier } from "domain/shared/identifier";
