--- 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/domain/exercise/value-object/exercise-group.ts
@@ -1,23 +1,0 @@
-import { ValueObject } from "domain/shared/value-object";
-
-export interface IExerciseGroup {
-  name: string;
-}
-
-export class ExerciseGroup extends ValueObject<IExerciseGroup> {
-  public get name(): string {
-    return this.props.name;
-  }
-
-  public static create(props: IExerciseGroup): ExerciseGroup {
-    if (props.name === "") {
-      throw new Error("Illegal name value.");
-    }
-
-    return new ExerciseGroup(props);
-  }
-
-  public equals(exerciseGroup: ExerciseGroup): boolean {
-    return this.props.name === exerciseGroup.name;
-  }
-}
